chore: make options sections more consistent
